backend/auth-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.6.0
pydantic-settings==2.1.0
fastapi-mail==1.4.1
python-dotenv==1.0.1
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
email-validator==2.1.0.post1